requirements.txt
aiohttp==3.10.10
beautifulsoup4==4.12.3
python-slugify==8.0.4